+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes
  - Default MFR class uses cho_factor and cho_solve from scipy
  - reworked format of line of sights, updated gmat computation functions
  - line of sight saving to a single json file
+ - new function for calculation of anisotropic derivative matrix
 
 New Features
 ------------
@@ -13,6 +14,7 @@ New Features
 
 Fixes
 -----
+ - updated derivative matrix checker
 
 Version 1.3 (21.07.2023)
 =======================